moved request function to separate file and send ids array for hanlers like params
